+++ webext/CHANGELOG.md
@@ -18,6 +18,7 @@ isn't!).
 ### Changed
 
 - Catch exception when trying to get securityInfo and the API doesn't exist.
+- Store and display all AltSvc headers we get, not just the SAT and onion ones
 
 ## v1.0.2 - 2019-02-19